+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - Added support for `DatetimeIndex.month_name` and `DatetimeIndex.day_name`.
 - Added support for `Series.dt.weekday`, `Series.dt.time`, and `DatetimeIndex.time`.
 - Added support for subtracting two timestamps to get a Timedelta.
+- Added support for `Series.dt.normalize`, and `DatetimeIndex.normalize`.
 
 #### Bug Fixes